+++ src/game/default/g_ai_node.c
@@ -21,7 +21,6 @@
 
 #include "g_local.h"
 #include "bg_pmove.h"
-#include "grid.h"
 
 /**
  * @brief Cached spatial acceleration structures for the navigation graph.
@@ -1434,7 +1433,7 @@ GArray *G_Ai_Node_FindPath(const g_client_t *cl, const ai_node_id_t start, const
 
   // size on 64k nodes is, say, 8kb.
   uint32_t costs_started[g_ai_nodes->len / 32 + 1] = {};
-  size_t visited = 0;
+  uint32_t visited = 0;
   // Min-heap open set (priority = f-cost). Replaces the previous sorted-array
   // queue (O(n) insertion) with an O(log n) binary heap from grid_ds.c.
   // Capacity is bounded by the node count plus headroom for re-insertions
@@ -1582,7 +1581,7 @@ GArray *G_Ai_Node_FindPath(const g_client_t *cl, const ai_node_id_t start, const
   GArray *return_path = NULL;
 
   if (finished) {
-    G_Ai_Debug("Found path from %u -> %u with %zu nodes visited\n", start, end, visited);
+    G_Ai_Debug("Found path from %u -> %u with %u nodes visited\n", start, end, visited);
 
     return_path = g_array_sized_new(false, false, sizeof(ai_node_id_t), visited);
